refactor: restructure movie module and routes, implement individual controller files
